backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-modeladmin-reorder==0.3.1
django-phonenumber-field==8.0.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
phonenumbers==8.13.55
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
